addTask function
get the value the user enters in the input.
If the user enters in nothing then it will not be added to the list.
we create a list element for the task. This element is given a class. The task is stored in the taskText Content and it is appended to the list.
The input value is then cleared